requirements.txt
nltk==3.6.5
python-Levenshtein==0.12.2
Flask==2.0.1
